refactor(post): 将帖子相关DTO中的数据类型从double修改为Long
scope: module/post/pojo/dto

将ListPostsResponse和SummaryPost类中涉及的字段数据类型从double更改为Long，以确保数据类型的准确性和一致性。

+++ src/main/java/top/xcyyds/wxfbackendclient/module/post/pojo/dto/ListPostsResponse.java
@@ -21,7 +21,7 @@ public class ListPostsResponse {
     /**
      * 当前页的帖子数量
      */
-    private double pageSize;
+    private Long pageSize;
     /**
      * 帖子列表
      */
@@ -29,6 +29,6 @@ public class ListPostsResponse {
     /**
      * 总帖子数
      */
-    private double totalPosts;
+    private Long totalPosts;
 }
